refactor: separate trade manager core and service
