+++ bot/trade_manager/__init__.py
@@ -1,0 +1,4 @@
+from .core import TradeManager
+from .service import api_app, asgi_app, create_trade_manager
+
+__all__ = ["TradeManager", "api_app", "asgi_app", "create_trade_manager"]
