+++ tests/test_trade_manager_loops.py
@@ -63,8 +63,8 @@ def _set_test_mode():
 @pytest.fixture(scope="module", autouse=True)
 def _import_trade_manager(_set_test_mode):
     global trade_manager, TradeManager
-    from bot import trade_manager as tm
-    from bot.trade_manager import TradeManager as TM
+    import bot.trade_manager.core as tm
+    from bot.trade_manager.core import TradeManager as TM
     trade_manager = tm
     TradeManager = TM
     yield
